This works to send a manual trace to the log
and fuck these deps were hard to figure out.
You have to include the exporters separately, and you can let the bom determine the versions.
But don't use the old exporters that have an s, use the opentelemetry-exporter-otlp etc.
And then it requires a grpc dep, and netty-shaded barfs, but grpc-netty worked.